requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
pymongo==4.4.1
redis==4.5.5
celery==5.3.1
pika==1.3.2
jinja2==3.1.2
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.2.0
factory-boy==3.2.1